requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# awkward π-nguin/test/stdout_server.py: 6
Click == 7.0

# awkward π-nguin/register_agent_service.py: 8
# awkward π-nguin/test/echo_server.py: 3
# awkward π-nguin/test/stdin_server.py: 3
# awkward π-nguin/test/stdout_server.py: 3
Flask == 1.1.1

# awkward π-nguin/baseagent.py: 27
PyXF == 1.0.6

# awkward π-nguin/baseagent.py: 32
# awkward π-nguin/registrar.py: 16
PyYAML == 3.13

# awkward π-nguin/debug.py: 13
aiodebug == 1.1.2

# awkward π-nguin/baseagent.py: 19
aiofiles == 0.6.0

# awkward π-nguin/baseagent.py: 20
aiohttp == 3.5.4

# awkward π-nguin/APiLexer.py: 2
# awkward π-nguin/APiListener.py: 2
# awkward π-nguin/APiParser.py: 3,809,822,835,848,859
# awkward π-nguin/listener.py: 6
antlr4_python3_runtime == 4.9

# awkward π-nguin/helpers.py: 4
fnvhash == 0.1.0

# awkward π-nguin/baseagent.py: 25
nclib == 1.0.0

# awkward π-nguin/baseagent.py: 26
pexpect == 4.6.0

# awkward π-nguin/baseagent.py: 44
psutil == 5.5.1

# awkward π-nguin/baseagent.py: 12,14
# awkward π-nguin/registrar.py: 6,8
requests == 2.22.0

# awkward π-nguin/baseagent.py: 8,9,10
spade == 3.1.4

# awkward π-nguin/baseagent.py: 24
# awkward π-nguin/test/stdin_ws_server.py: 3
# awkward π-nguin/test/stdin_ws_server_3682.py: 3
# awkward π-nguin/test/ws_echo_server.py: 3
# awkward π-nguin/test/ws_stdout_server.py: 3
websockets == 8.1
